requirements.txt
python-telegram-bot==20.3
cryptography==41.0.3
octodiary==0.3.0
